Move source files and assign pki.server.healthcheck.* namespace
- Move source files to conform with convention
- Assign pki.server.healthcheck.* namespace to the healthcheck tool
- Unassign pki.server.healthcheck as a namespace package (creates conflict
  with pki.* packages)
