+++ base/server/healthcheck/CMakeLists.txt
@@ -9,11 +9,16 @@ set(PYTHON "python3")
 
 set(SETUP_PY_IN "${CMAKE_CURRENT_SOURCE_DIR}/setup.py")
 set(SETUP_PY    "${CMAKE_CURRENT_BINARY_DIR}/setup.py")
-set(SRC_DIR     "pkihealthcheck")
+set(SRC_DIR     "pki")
 set(INSTALL_DIR "${CMAKE_CURRENT_BINARY_DIR}/install")
 
 configure_file(${SETUP_PY_IN} ${SETUP_PY})
-configure_file(${CMAKE_CURRENT_SOURCE_DIR}/pki-healthcheck.conf ${CMAKE_CURRENT_BINARY_DIR}/pki-healthcheck.conf @ONLY)
+
+# Customize default healthcheck.conf
+configure_file(
+    ${CMAKE_CURRENT_SOURCE_DIR}/healthcheck.conf
+    ${CMAKE_CURRENT_BINARY_DIR}/healthcheck.conf @ONLY
+)
 
 add_custom_target(${target} ALL
     DEPENDS ${SETUP_PY}
@@ -53,7 +58,8 @@ install(
 
 install(
     FILES
-        ${CMAKE_CURRENT_BINARY_DIR}/pki-healthcheck.conf
+        ${CMAKE_CURRENT_BINARY_DIR}/healthcheck.conf
     DESTINATION
         ${SYSCONF_INSTALL_DIR}/pki/
 )
+
