+++ base/server/healthcheck/setup.py
@@ -1,30 +1,31 @@
-from setuptools import find_packages, setup
+from setuptools import setup
 
 
 setup(
+    # A hyphen ('-') gets converted to underscore ('_') while packaging
+    # so avoiding the name pki-healthcheck
     name='pkihealthcheck',
     version='0.1',
-    namespace_packages=['pkihealthcheck'],
     packages=[
-        'pkihealthcheck.core',
-        'pkihealthcheck.pki',
+        'pki.server.healthcheck.core',
+        'pki.server.healthcheck.meta',
     ],
     entry_points={
         # creates bin/pki-healthcheck
         'console_scripts': [
-            'pki-healthcheck = pkihealthcheck.core.main:main',
+            'pki-healthcheck = pki.server.healthcheck.core.main:main'
         ],
         # register the plugin with ipa-healthcheck
         'ipahealthcheck.registry': [
-            'pkihealthcheck.pki = pkihealthcheck.pki.plugin:registry',
+            'pkihealthcheck.pki = pki.server.healthcheck.meta.plugin:registry',
         ],
         # register the plugin with pki-healthcheck
         'pkihealthcheck.registry': [
-            'pkihealthcheck.pki = pkihealthcheck.pki.plugin:registry',
+            'pkihealthcheck.pki = pki.server.healthcheck.meta.plugin:registry',
         ],
         # plugin modules for pkihealthcheck.pki registry
         'pkihealthcheck.pki': [
-            'pki_certs = pkihealthcheck.pki.certs',
+            'pki_certs = pki.server.healthcheck.meta.csconfig',
         ],
     },
     classifiers=[
